requirements.txt
datasets>=1.11.0
flatten-dict
hydra-core>=1.0.6
isort==5.9.3
librosa
pycocotools
pydicom==2.3.0
pylint==2.11.0
pytorch-lightning==1.3.0
scipy
toml
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.6.0
torchvision==0.14.0
tqdm==4.64.1
transformers
wandb==0.13.4
wget
yapf
einops
hub
wfdb
timm
h5py
pylidc
bs4
mne
